removed InitializeSubscription hook + added SubscriptionContext to handle customer, products
