+++ client/packages/lowcoder/src/redux/sagas/subscriptionSagas.ts
@@ -1,12 +1,12 @@
 import { call, put, select, takeLatest, all, take } from 'redux-saga/effects';
 import { fetchSubscriptionsAction } from 'redux/reduxActions/subscriptionActions';
-import { searchCustomersSubscriptions, Subscription } from 'api/subscriptionApi';
+import { searchCustomersSubscriptions } from 'api/subscriptionApi';
 import { fetchSubscriptionsSuccess, fetchSubscriptionsError } from 'redux/reduxActions/subscriptionActions';
-import { LowcoderSearchCustomer } from 'api/subscriptionApi';
 import { getUser, getCurrentUser } from 'redux/selectors/usersSelectors';
 import { getDeploymentId } from "redux/selectors/configSelectors";
 import { CurrentUser, User } from '@lowcoder-ee/constants/userConstants';
 import { ReduxActionTypes } from '@lowcoder-ee/constants/reduxActionConstants';
+import { Subscription, LowcoderSearchCustomer } from '@lowcoder-ee/constants/subscriptionConstants';
 
 function* fetchSubscriptionsSaga(action: ReturnType<typeof fetchSubscriptionsAction>) {
   try {
