+++ client/packages/lowcoder/src/util/hooks.ts
@@ -9,7 +9,7 @@ import React, {
   useRef,
   useState,
 } from "react";
-import { shallowEqual, useSelector } from "react-redux";
+import { shallowEqual, useDispatch, useSelector } from "react-redux";
 import { useLocation, useParams } from "react-router-dom";
 import { DATASOURCE_URL, QUERY_LIBRARY_URL } from "../constants/routesURL";
 import { AuthSearchParams } from "constants/authConstants";
@@ -26,6 +26,9 @@ import { CompAction, changeChildAction } from "lowcoder-core";
 import { ThemeDetail } from "@lowcoder-ee/api/commonSettingApi";
 import { uniq } from "lodash";
 import { constantColors } from "components/colorSelect/colorUtils";
+import { AppState } from "@lowcoder-ee/redux/reducers";
+import { getOrgUsers } from "@lowcoder-ee/redux/selectors/orgSelectors";
+import { fetchOrgUsersAction } from "@lowcoder-ee/redux/reduxActions/orgActions";
 
 export const ForceViewModeContext = React.createContext<boolean>(false);
 
@@ -257,3 +260,25 @@ export function useThemeColors(allowGradient?: boolean) {
     defaultTheme,
   ]);
 }
+
+export const useOrgUserCount = (orgId: string) => {
+  const dispatch = useDispatch();
+  const orgUsers = useSelector((state: AppState) => getOrgUsers(state)); // Use selector to get orgUsers from state
+  const [userCount, setUserCount] = useState<number>(0);
+
+  useEffect(() => {
+    // Dispatch action to fetch organization users
+    if (orgId) {
+      dispatch(fetchOrgUsersAction(orgId));
+    }
+  }, [dispatch, orgId]);
+
+  useEffect(() => {
+    // Update user count when orgUsers state changes
+    if (orgUsers && orgUsers.length > 0) {
+      setUserCount(orgUsers.length);
+    }
+  }, [orgUsers]);
+
+  return userCount;
+};
